Add: optimize trigger
It is NOT a good idea to disable spawn for Helicopter because city will
not be activate until player get out. The AI will spawn directly infront
of players ...

A clever way could be to activate city when the pilot start to decrease
the helicopter speed for landing. So during the travel, no AI will spawn
until you start to apporch your landing zone.