FIX: add a delay for ownership broadcast
The owner command takes some times to transfer ownership to client.
Setmass doesn't work during the transfer of ownership .  So this delay
is necessary because the setmass command need to be apply after the
broadcast and not DURING the transfer of ownership. Due to this delay
the btc_lifted need to be set as soon as possible to limit multiple
spawn of lift_hook function.